优化订单号规则

+++ src/main/java/com/jkf/channel/gateway/dao/SysConfigMapper.java
@@ -33,5 +33,5 @@ public interface SysConfigMapper {
      * @param seqName
      * @return
      */
-    int getSeq(String seqName);
+    long getSeq(String seqName);
 }
